requirements.txt
discord.py
re.py